requirements_exact.txt
alembic==1.18.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
hydra-core==1.3.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
optuna==4.7.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.4
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentry-sdk==2.50.0
setuptools==80.10.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.1
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.24.0